Fix phpseclib for PHP 8.2

+++ scripts/phpseclib/Net/SSH2.php
@@ -157,6 +157,7 @@ define('NET_SSH2_READ_NEXT', 3);
  * @author  Jim Wigginton <terrafrost@php.net>
  * @access  public
  */
+#[AllowDynamicProperties]
 class Net_SSH2
 {
     /**
@@ -1217,8 +1218,8 @@ class Net_SSH2
                 $read = array($this->fsock);
                 $write = $except = null;
                 $start = strtok(microtime(), ' ') + strtok('');
-                $sec = floor($this->curTimeout);
-                $usec = 1000000 * ($this->curTimeout - $sec);
+                $sec = (int) floor($this->curTimeout);
+                $usec = (int) (1000000 * ($this->curTimeout - $sec));
                 // on windows this returns a "Warning: Invalid CRT parameters detected" error
                 // the !count() is done as a workaround for <https://bugs.php.net/42682>
                 if (!@stream_select($read, $write, $except, $sec, $usec) && !count($read)) {
@@ -3810,8 +3811,8 @@ class Net_SSH2
                     $write = $except = null;
 
                     $start = strtok(microtime(), ' ') + strtok(''); // http://php.net/microtime#61838
-                    $sec = floor($this->curTimeout);
-                    $usec = 1000000 * ($this->curTimeout - $sec);
+                    $sec = (int) floor($this->curTimeout);
+                    $usec = (int) (1000000 * ($this->curTimeout - $sec));
                     // on windows this returns a "Warning: Invalid CRT parameters detected" error
                     if (!@stream_select($read, $write, $except, $sec, $usec) && !count($read)) {
                         $this->is_timeout = true;
